Move RhoCmds to their own files (#52)
* Use Ansible variables, not dynamic role files

Move towards using JBoss Scanner's facts inside of Rho by using static
Ansible role files with variables. (Previous Rho was dynamically
generating the role files at runtime.) This will make it much easier
to use a larger number of tasks in the roles, which is important
because JBoss Scanner uses many Ansible tasks.

* Use Ansible variables, not dynamic role files

Move towards using JBoss Scanner's facts inside of Rho by using static
Ansible role files with variables. (Previous Rho was dynamically
generating the role files at runtime.) This will make it much easier
to use a larger number of tasks in the roles, which is important
because JBoss Scanner uses many Ansible tasks.

* Fix lint problems.

* Move RhoCmds to their own files

This only affects the Python module structure of the program, not the
Ansible modules. However, this change is a prerequisite for changing
the Ansible module structure, if we want to do that, and also makes
the code cleaner and easier to read.

I've tried to keep this PR completely mechanical - just move code from
run_cmds.py to other files and update names and include statements
appropriately.

* Fix linter warnings
